fix: address PR #187 review — floor overflow, stale wedge, diagnostics wiring
- Floor math in i64 at all four sites (config load + identity classifier
  + identity validate): keep_versions is unbounded operator input, so
  the u32 add could overflow and the as-i32 cast could wrap negative,
  silently disabling the blocks_to_keep floor. Regression test at
  keep_versions = u32::MAX.
- Clear the deep-fork wedge as soon as the fork point is reachable
  (ForkPoint::Found), not only after a successful rollback — a transient
  rollback_to failure no longer leaves /health stuck on a stale wedge.
- /diagnostics/fork-risk now populates fork_too_deep / resync_required
  from ApiStatus.sync_wedged (both were hardcoded None with an
  'ASSUMED-new' unknown-vector note that this PR made stale).
- Derive Default on ApiStatus + SyncStateLabel (ApiIdentity pattern) so
  future field additions stop breaking every test stub; stale
  ROLLBACK_WINDOW wording in the resolved-config doc updated to the
  configurable keep_versions floor. OpenAPI golden regenerated.

+++ ergo-node/src/config/load.rs
@@ -378,8 +378,11 @@ impl NodeConfig {
         // Mode 6 (`blocks_to_keep = 0`) is exempt — it's headers-only
         // and never applies full blocks.
         if blocks_to_keep > 0 {
-            let floor = (keep_versions + ergo_state::store::SAFETY_MARGIN) as i32;
-            if blocks_to_keep < floor {
+            // i64 math: `keep_versions` is unbounded operator input, so a
+            // u32 add can overflow and an `as i32` cast can wrap negative —
+            // silently disabling this floor for huge windows.
+            let floor = i64::from(keep_versions) + i64::from(ergo_state::store::SAFETY_MARGIN);
+            if i64::from(blocks_to_keep) < floor {
                 return Err(format!(
                     "[node] blocks_to_keep = {blocks_to_keep} is below the rollback-window \
                      floor ({floor} = keep_versions {keep_versions} + SAFETY_MARGIN {}). \
